requirements.txt
python-telegram-bot>=21.0
pymongo>=4.10.0
dnspython>=2.6.0
